fix cart/orders bug after not processed order, add cart item quantity update

+++ order-service/src/main/java/com/gritlab/buy01/orderservice/service/OrderService.java
@@ -94,9 +94,11 @@ public class OrderService {
       cartService.updateCartContents(modifications.getModifications());
     }
 
-    for (Order order : cartValidationResponse.getCart().getOrders()) {
-      orderRepository.save(order);
-      cartService.deleteItemFromCart(order.getId(), userId);
+    if (cartValidationResponse.getProcessed()) {
+      for (Order order : cartValidationResponse.getCart().getOrders()) {
+        orderRepository.save(order);
+        cartService.deleteItemFromCart(order.getId(), userId);
+      }
     }
 
     return new CartResponse(
